Refactor the database framerate limiter with a dedicated module
