+++ src/database/mod.rs
@@ -1,7 +1,9 @@
 mod execution;
+mod framerate;
 mod query;
 mod storage;
 
+use framerate::Clock;
 use chrono::offset::Utc;
 use crate::execution::{Request as ExecutionRequest, Response as ExecutionResponse};
 use crate::query::{Request as QueryRequest, Response as QueryResponse};
@@ -10,7 +12,6 @@ use self::query::Handler as QueryHandler;
 use self::storage::Storage;
 use std::sync::mpsc::{Receiver, Sender};
 use std::thread;
-use std::time::{Duration, Instant};
 
 /// Start the Database, spawning a thread and returning the join handle.
 ///
@@ -18,7 +19,7 @@ use std::time::{Duration, Instant};
 /// receive queries from the Controller, to send query responses to the Controller, to send
 /// execution requests to the Processor, and to receive execution responses from the Processor. The
 /// Database runs in its own process, at its own framerate. The job and rule storage is delegated
-/// to an Storage. The Storage is initialized at the start of the Database process.
+/// to a [`Storage`]. The [`Storage`] is initialized at the start of the Database process.
 pub fn start(
     query_link: (Sender<QueryResponse>, Receiver<QueryRequest>),
     execution_link: (Sender<ExecutionRequest>, Receiver<ExecutionResponse>),
@@ -32,23 +33,11 @@ pub fn start(
             panic!("Unable to initialize the storage from data persisted to the file system.");
         };
 
-        loop {
-            let previous_time = Instant::now();
-
+        let clock = Clock::with_framerate(128);
+        clock.start(|| {
             let current_datetime = Utc::now();
             query_handler.handle(&current_datetime, &mut storage);
             execution_handler.handle(&current_datetime, &mut storage);
-
-            // Put the thread asleep to run at a maximum of 128 time per second.
-            let now = Instant::now();
-            let elapsed_time = now.duration_since(previous_time);
-
-            match Duration::new(0, 1_000_000_000u32 / 128).checked_sub(elapsed_time) {
-                Some(sleep_time) => {
-                    thread::sleep(sleep_time);
-                },
-                None => {},
-            };
-        };
+        });
     })
 }
